+++ packages/shared/src/index.ts
@@ -1,6 +1,6 @@
-import { MapSchema, Schema, type } from "@colyseus/schema";
-
 export * from "./spawn-room";
+export * from "./messages";
+export * from "./state";
 
 export const ROOM_NAME = "game";
 export const DEFAULT_CHARACTER_NAME = "Player";
@@ -10,35 +10,3 @@ export const SEND_DELTA_TIME = 1000 / SEND_RATE;
 export type JoinOptions = {
   name?: string;
 };
-
-export namespace SetTransform {
-  export const Type = "set-transform";
-
-  export type Message = {
-    x: number;
-    y: number;
-    z: number;
-  };
-}
-
-export class PositionState extends Schema {
-  @type("number") x: number = 0;
-  @type("number") y: number = 0;
-  @type("number") z: number = 0;
-}
-
-export class RotationState extends Schema {
-  @type("number") x: number = 0;
-  @type("number") y: number = 0;
-  @type("number") z: number = 0;
-}
-
-export class CharacterState extends Schema {
-  @type("string") name: string = "Player";
-  @type(PositionState) position: PositionState = new PositionState();
-  @type(RotationState) rotation: RotationState = new RotationState();
-}
-
-export class GameState extends Schema {
-  @type({ map: CharacterState }) characters = new MapSchema<CharacterState>();
-}
