+++ packages/shared/src/messages.ts
@@ -1,0 +1,9 @@
+export namespace SetTransform {
+  export const Type = "set-transform";
+
+  export type Message = {
+    x: number;
+    y: number;
+    z: number;
+  };
+}
