+++ packages/shared/src/state.ts
@@ -1,0 +1,23 @@
+import { MapSchema, Schema, type } from "@colyseus/schema";
+
+export class PositionState extends Schema {
+  @type("number") x: number = 0;
+  @type("number") y: number = 0;
+  @type("number") z: number = 0;
+}
+
+export class RotationState extends Schema {
+  @type("number") x: number = 0;
+  @type("number") y: number = 0;
+  @type("number") z: number = 0;
+}
+
+export class CharacterState extends Schema {
+  @type("string") name: string = "Player";
+  @type(PositionState) position: PositionState = new PositionState();
+  @type(RotationState) rotation: RotationState = new RotationState();
+}
+
+export class GameState extends Schema {
+  @type({ map: CharacterState }) characters = new MapSchema<CharacterState>();
+}
